Move cookieauth plugin lock implementation to base
