+++ plugins/base.js
@@ -13,6 +13,12 @@
 // limitations under the License.
 'use strict';
 
+const debug = require('debug')('cloudant:plugins:base');
+const lockFile = require('lockfile');
+const tmp = require('tmp');
+
+function noop() {}
+
 /**
  * Cloudant base plugin.
  *
@@ -22,6 +28,32 @@
 class BasePlugin {
   constructor(client, cfg) {
     this._client = client;
+    this._lockFile = tmp.tmpNameSync({ postfix: '.lock' });
+  }
+
+  // Acquire a file lock on the specified path. Release the file lock on
+  // completion of the callback.
+  withLock(opts, callback) {
+    var self = this;
+    if (typeof opts === 'function') {
+      callback = opts;
+      opts = {};
+    }
+    debug('Acquiring lock...');
+    lockFile.lock(self._lockFile, opts, function(error) {
+      if (error) {
+        callback(error, noop);
+      } else {
+        callback(null, function() {
+          // Close and unlink the file lock.
+          lockFile.unlock(self._lockFile, function(error) {
+            if (error) {
+              debug(`Failed to release lock: ${error}`);
+            }
+          });
+        });
+      }
+    });
   }
 
   get id() {
